Use build tags for legacy plugin
