+++ provider/build_config_default.go
@@ -1,0 +1,6 @@
+//go:build !legacy_plugin
+
+package provider
+
+// useLegacyProvider indicates whether to use the legacy hashicorp/go-plugin implementation
+const useLegacyProvider = false
